% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardioCTGroundTruth}
\alias{cardioCTGroundTruth}
\title{{
Ground Truth for the Cardiac CT study
}}
\format{A data frame with 210 rows and 2 variables:
\describe{
  \item{caseID}{ID for 210 cases, among which 105 positive for lesion (1--209)}
  \item{truth}{factor with two levels: positive(with lesion) / negative
   (without lesion)}
}}
\usage{
cardioCTGroundTruth
}
\description{
{
This truth data was produced from the color scale study for the Cardiac CT
 images described in

 \emph{Zabala-Travers S, Gallas BD, Busoni S, Williams MC, Noferini L, Fedeli L,
 Lucarini S, Galastri L, Mirsadraee S, Badano A. Display colour scale
 effects on diagnostic performance and reader agreement in cardiac CT and
 prostate apparent diffusion coefficient assessment. Clinical radiology.
 2019 Jan 1;74(1):79-e1.}

 The variables are as follows:
}
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/30336942}
}
\keyword{datasets}
