% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardioCTReaderData}
\alias{cardioCTReaderData}
\title{{
Reader study result for the Cardiac CT study
}}
\format{A data frame with 1921 rows and 4 variables:
\describe{
  \item{readerID}{Readers(obervers) had different experience level, which
   is referenced in the reader nickname as High and Low, for high and low
   level of experience, respectively. (High1--High4, Low1--Low8)}
  \item{caseID}{ID for 210 cases, among which 105 positive for lesion (1--210)}
  \item{modalityID}{Three color scales: Grayscale, Rainbow, Hotiron(Hot).
   Grayscale was evaluated using GSDF settings and Rainbow and Hotiron using
   RGB settings. }
  \item{confidenceScore}{The score gave from the readers (0--100)}
}}
\usage{
cardioCTReaderData
}
\description{
{
This reader study result was produced from the color scale study for the
 Cardiac CT images described in

 \emph{Zabala-Travers S, Gallas BD, Busoni S, Williams MC, Noferini L, Fedeli L,
 Lucarini S, Galastri L, Mirsadraee S, Badano A. Display colour scale
 effects on diagnostic performance and reader agreement in cardiac CT and
 prostate apparent diffusion coefficient assessment. Clinical radiology.
 2019 Jan 1;74(1):79-e1.}

 The study is not fully crossed. The variables are as follows:
}
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/30336942}
}
\keyword{datasets}
