% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostateMRIGroundTruth}
\alias{prostateMRIGroundTruth}
\title{{
Ground Truth for the Prostate MRI study
}}
\format{A data frame with 165 rows and 2 variables:
\describe{
  \item{caseID}{ID for 165 cases, among which 66 positive for lesion (1--165)}
  \item{truth}{factor with two levels: positive(with lesion) / negative
   (without lesion)}
}}
\usage{
prostateMRIGroundTruth
}
\description{
{
This truth data was produced from the color scale study for the
 Prostate MRI images described in

 \emph{Zabala-Travers S, Gallas BD, Busoni S, Williams MC, Noferini L, Fedeli L,
 Lucarini S, Galastri L, Mirsadraee S, Badano A. Display colour scale
 effects on diagnostic performance and reader agreement in cardiac CT and
 prostate apparent diffusion coefficient assessment. Clinical radiology.
 2019 Jan 1;74(1):79-e1.}

 The variables are as follows:
}
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/30336942}
}
\keyword{datasets}
