% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostateMRIReaderData}
\alias{prostateMRIReaderData}
\title{{
Reader study result for the Prostate MRI study
}}
\format{A data frame with 2035 rows and 4 variables:
\describe{
  \item{readerID}{Readers(obervers) had different experience level, which
   is referenced in the reader nickname as High, Med and Low, for high, medium
   and low level of experience, respectively. (High1--High3, Med1--Med3, Low1
   --Low3)}
  \item{caseID}{ID for 165 cases, among which 66 positive for lesion (1--165)}
  \item{modalityID}{Five color scales: Grayscale, Rainbow, Hotiron, Rainbow
   read with GSDF, Hotiron read with GSDF }
  \item{confidenceScore}{The score gave from the readers (0--100)}
}}
\usage{
prostateMRIReaderData
}
\description{
{
This reader study result was produced from the color scale study for the
 Prostate MRI images described in

 \emph{Zabala-Travers S, Gallas BD, Busoni S, Williams MC, Noferini L, Fedeli L,
 Lucarini S, Galastri L, Mirsadraee S, Badano A. Display colour scale
 effects on diagnostic performance and reader agreement in cardiac CT and
 prostate apparent diffusion coefficient assessment. Clinical radiology.
 2019 Jan 1;74(1):79-e1.}

 The study is not fully crossed. The variables are as follows:
}
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/30336942}
}
\keyword{datasets}
